requirements.txt
requests
psycopg2
pyyaml
kafka-python
codecov